Changed UserMed data type to integer

+++ src/main/java/medtrackercapstone/medtracker/database/dao/UserMedDAO.java
@@ -19,7 +19,7 @@ import java.util.List;
 
 
 
-public interface UserMedDAO extends JpaRepository<UserMed, Long> {
+public interface UserMedDAO extends JpaRepository<UserMed, Integer> {
 
 //    UserMed findById(@Param("id") Integer id);
